파이어스토어 데이터 거리 기준으로 수정

+++ lib/logic/map/market_add_data/map_marker_data.dart
@@ -19,7 +19,7 @@ sealed class MarketData with _$MarketData{
 
     double? gpsX,
     double? gpsY,
-    int? distance,
+    double? distance,
   }) = _MarketData;
   factory MarketData.fromJson(Map<String,dynamic> json ) => _$MarketDataFromJson(json);
 }
